backend/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
Faker==12.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.7.0
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
Faker==12.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.7.0
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
Faker==12.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
zipp==3.7.0
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.3
decorator==5.1.1
Django==4.0.2
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.2
Faker==12.0.0
importlib-metadata==4.10.1
ipython==8.0.1
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
zipp==3.7.0
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.3
decorator==5.1.1
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
executing==0.8.2
Faker==12.0.0
importlib-metadata==4.10.1
ipython==8.0.1
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
zipp==3.7.0
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.3
decorator==5.1.1
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
executing==0.8.2
Faker==12.0.0
importlib-metadata==4.10.1
ipython==8.0.1
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
tomli==2.0.0
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
zipp==3.7.0
